# contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/work-orders": {
      "post": {
        "operationId": "WorkOrderController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkOrderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request body is invalid."
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a work order",
        "tags": [
          "work-orders"
        ]
      },
      "get": {
        "operationId": "WorkOrderController_findAll_v1",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkOrderResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List work orders",
        "tags": [
          "work-orders"
        ]
      }
    },
    "/api/v1/work-orders/{workOrderId}/extra-expenses": {
      "post": {
        "operationId": "WorkOrderController_createExtraExpense_v1",
        "parameters": [
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExtraExpenseRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtraExpenseResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request body is invalid."
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create an extra expense",
        "tags": [
          "work-orders"
        ]
      },
      "get": {
        "operationId": "WorkOrderController_findAllExtraExpenses_v1",
        "parameters": [
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExtraExpenseResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List extra expenses",
        "tags": [
          "work-orders"
        ]
      }
    },
    "/api/v1/work-orders/{workOrderId}/history": {
      "post": {
        "operationId": "WorkOrderController_createHistory_v1",
        "parameters": [
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkOrderHistoryRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderHistoryItemResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request body is invalid."
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a work order history entry",
        "tags": [
          "work-orders"
        ]
      },
      "get": {
        "operationId": "WorkOrderController_findAllHistory_v1",
        "parameters": [
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkOrderHistoryItemResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List work order history",
        "tags": [
          "work-orders"
        ]
      }
    },
    "/api/v1/work-orders/{workOrderId}": {
      "get": {
        "operationId": "WorkOrderController_findById_v1",
        "parameters": [
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a work order",
        "tags": [
          "work-orders"
        ]
      },
      "patch": {
        "operationId": "WorkOrderController_update_v1",
        "parameters": [
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkOrderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request body is invalid."
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a work order",
        "tags": [
          "work-orders"
        ]
      },
      "delete": {
        "operationId": "WorkOrderController_delete_v1",
        "parameters": [
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a work order",
        "tags": [
          "work-orders"
        ]
      }
    },
    "/api/v1/work-orders/{workOrderId}/extra-expenses/{extraExpenseId}": {
      "get": {
        "operationId": "WorkOrderController_findExtraExpenseById_v1",
        "parameters": [
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "extraExpenseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtraExpenseResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get an extra expense",
        "tags": [
          "work-orders"
        ]
      },
      "patch": {
        "operationId": "WorkOrderController_updateExtraExpense_v1",
        "parameters": [
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "extraExpenseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExtraExpenseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtraExpenseResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request body is invalid."
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update an extra expense",
        "tags": [
          "work-orders"
        ]
      },
      "delete": {
        "operationId": "WorkOrderController_deleteExtraExpense_v1",
        "parameters": [
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "extraExpenseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete an extra expense",
        "tags": [
          "work-orders"
        ]
      }
    }
  },
  "info": {
    "title": "Faberon API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "message": {
            "type": "string",
            "example": "Resource was not found."
          },
          "error": {
            "type": "string",
            "example": "Not Found"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "CreateWorkOrderRequest": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "description": {
            "type": "string",
            "example": "Paper jams repeatedly.",
            "minLength": 1,
            "maxLength": 4000
          },
          "productId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "workerId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "estimatedDate": {
            "type": "string",
            "format": "date",
            "example": "2026-07-24",
            "nullable": true
          },
          "showFinalPrice": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "customerId",
          "description"
        ]
      },
      "WorkOrderResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "companyId": {
            "type": "string",
            "format": "uuid"
          },
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "workerId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "number": {
            "type": "number",
            "example": 1042
          },
          "description": {
            "type": "string"
          },
          "stage": {
            "type": "string",
            "enum": [
              "waiting",
              "diagnostics",
              "waiting-parts",
              "repaired"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "estimatedDate": {
            "type": "string",
            "nullable": true,
            "format": "date"
          },
          "showFinalPrice": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "companyId",
          "customerId",
          "productId",
          "workerId",
          "number",
          "description",
          "stage",
          "createdAt",
          "estimatedDate",
          "showFinalPrice"
        ]
      },
      "CreateExtraExpenseRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Fuser assembly",
            "minLength": 1,
            "maxLength": 200
          },
          "price": {
            "type": "string",
            "example": "210.00"
          },
          "isHidden": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "name",
          "price"
        ]
      },
      "ExtraExpenseResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "workOrderId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "string",
            "example": "210.00"
          },
          "isHidden": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "workOrderId",
          "name",
          "price",
          "isHidden"
        ]
      },
      "CreateWorkOrderHistoryRequest": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "example": "Assigned to self.",
            "minLength": 1,
            "maxLength": 4000
          }
        },
        "required": [
          "text"
        ]
      },
      "WorkOrderHistoryItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "workOrderId": {
            "type": "string",
            "format": "uuid"
          },
          "workerId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "text": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "workOrderId",
          "workerId",
          "text",
          "createdAt"
        ]
      },
      "UpdateWorkOrderRequest": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "description": {
            "type": "string",
            "example": "Paper jams repeatedly.",
            "minLength": 1,
            "maxLength": 4000
          },
          "productId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "workerId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "stage": {
            "type": "string",
            "enum": [
              "waiting",
              "diagnostics",
              "waiting-parts",
              "repaired"
            ]
          },
          "estimatedDate": {
            "type": "string",
            "format": "date",
            "example": "2026-07-24",
            "nullable": true
          },
          "showFinalPrice": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "UpdateExtraExpenseRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Fuser assembly",
            "minLength": 1,
            "maxLength": 200
          },
          "price": {
            "type": "string",
            "example": "210.00"
          },
          "isHidden": {
            "type": "boolean",
            "example": false
          }
        }
      }
    }
  }
}